+++ src/lib/Makefile.inc
@@ -21,6 +21,7 @@ CSOURCES = ares__addrinfo2hostent.c	\
   ares_cancel.c				\
   ares_data.c				\
   ares_destroy.c			\
+  ares_dns_mapping.c	\
   ares_expand_name.c			\
   ares_expand_string.c			\
   ares_fds.c				\
